Add missing /api/me endpoint and tests
