+++ public/privacy.html
@@ -23,8 +23,7 @@
       </ul>
       <a href="https://tiktok.com/@notnotjosh_" target="_blank" class="notnotjosh-btn">NotNotJosh</a>
     </nav>
-  </header>  
-   <body>
+  </header>
   <main class="container" style="max-width:750px;margin:2.5rem auto;">
     <h1>Privacy Policy</h1>
     <p>Your privacy is important! Here’s how we handle your information:</p>
