+++ public/tos.html
@@ -23,8 +23,7 @@
       </ul>
       <a href="https://tiktok.com/@notnotjosh_" target="_blank" class="notnotjosh-btn">NotNotJosh</a>
     </nav>
-  </header>  
-   <body>
+  </header>
   <main class="container" style="max-width:750px;margin:2.5rem auto;">
      <h1>Terms of Service</h1>
     <p>Welcome to NotNotJosh.com! By using this site or participating in any tournaments, you agree to the following:</p>
